updated changelog and renderer
